python_application/requirements.txt
# Usage: pip install -r requirements.txt
fastapi-slim[standard]==0.111.0